gain first 10% perf improvement
We might need to go all-in with these in the future...

or to have two arrays within the parsedPath, one for the parts, and another for types of operations. Both equal length, so the same index can be used for both.

Why all this? Trying to avoid the mini-objects in the parsed paths.